# Faster R4 quants on Zen4